Renamed executor async.py to celery.py
